fix clean script

+++ scripts/clean-env.js
@@ -15,15 +15,23 @@ async function main() {
         hostEcm: program.host
     });
 
-    await this.alfrescoJsApi.login(program.username, program.password);
+    try {
+        await this.alfrescoJsApi.login(program.username, program.password);
+    } catch (error) {
+        console.log(JSON.stringify(error));
+    }
 
+    console.log('====== Clean Root ======');
     await cleanRoot(this.alfrescoJsApi);
+
+    console.log('====== Clean Sites ======');
     await deleteSite(this.alfrescoJsApi);
+
+    console.log('====== Empty Trash ======');
     await emptyTrashCan(this.alfrescoJsApi);
 }
 
 async function cleanRoot(alfrescoJsApi) {
-    console.log('====== Clean Root ======');
 
     let rootNodes = await alfrescoJsApi.core.nodesApi.getNodeChildren('-root-', {
         include: ['properties']
@@ -33,7 +41,7 @@ async function cleanRoot(alfrescoJsApi) {
 
         sleep(200);
 
-        if(rootNodes.list.entries[i].entry.createdByUser.id !== 'System') {
+        if (rootNodes.list.entries[i].entry.createdByUser.id !== 'System') {
 
             try {
                 await alfrescoJsApi.core.nodesApi.deleteNode(rootNodes.list.entries[i].entry.id);
@@ -46,49 +54,48 @@ async function cleanRoot(alfrescoJsApi) {
 }
 
 async function emptyTrashCan(alfrescoJsApi) {
-    console.log('====== Clean Trash ======');
-
     let deletedNodes = await alfrescoJsApi.core.nodesApi.getDeletedNodes();
 
-    for (let i = 0; i < deletedNodes.list.entries.length; i++) {
+    if (deletedNodes.list.entries.length > 0) {
+        for (let i = 0; i < deletedNodes.list.entries.length; i++) {
 
-        sleep(200);
+            sleep(200);
 
-        console.log(deletedNodes.list.entries[i].entry.id);
+            console.log(deletedNodes.list.entries[i].entry.id);
 
-        try {
-            await alfrescoJsApi.core.nodesApi.purgeDeletedNode(deletedNodes.list.entries[i].entry.id);
-        } catch (error) {
-            console.log('error' + JSON.stringify(error));
+            try {
+                await alfrescoJsApi.core.nodesApi.purgeDeletedNode(deletedNodes.list.entries[i].entry.id);
+            } catch (error) {
+                console.log('error' + JSON.stringify(error));
+            }
         }
-    }
 
-    emptyTrashCan(alfrescoJsApi);
+        emptyTrashCan(alfrescoJsApi);
+    }
 }
 
 async function deleteSite(alfrescoJsApi) {
-    console.log('====== Clean Sites ======');
-
     let listSites = await this.alfrescoJsApi.core.sitesApi.getSites();
 
-    console.log(listSites.list.pagination.totalItems);
-    for (let i = 0; i < listSites.list.entries.length; i++) {
+    if (listSites.list.pagination.totalItems > 1) {
+        for (let i = 0; i < listSites.list.entries.length; i++) {
 
-        sleep(200);
+            sleep(200);
 
-        console.log(listSites.list.entries[i].entry.id);
+            console.log(listSites.list.entries[i].entry.id);
 
-        if (listSites.list.entries[i].entry.id !== 'swsdp') {
-            try {
-                await alfrescoJsApi.core.sitesApi.deleteSite(listSites.list.entries[i].entry.id, {options: {permanent: true}});
-            } catch (error) {
-                console.log('error' + JSON.stringify(error));
+            if (listSites.list.entries[i].entry.id !== 'swsdp') {
+                try {
+                    await alfrescoJsApi.core.sitesApi.deleteSite(listSites.list.entries[i].entry.id, {options: {permanent: true}});
+                } catch (error) {
+                    console.log('error' + JSON.stringify(error));
 
+                }
             }
         }
-    }
 
-    deleteSite(alfrescoJsApi);
+        deleteSite(alfrescoJsApi);
+    }
 }
 
 function sleep(delay) {
